refactored firewall page
