+++ src/app/api/instances/[name]/firewall/types.ts
@@ -1,0 +1,7 @@
+import { FirewallRule } from "@/types/firewall";
+
+export interface UpdateFireWallRulesParams {
+  region: string;
+  rules: FirewallRule[];
+  instanceName: string;
+}
